removed flot pos = (edited_df[['Hydrothermal...

+++ app.py
@@ -78,7 +78,7 @@ original_title = '<p style="font-family:Courier; color:Black; font-size: 30px;">
 st.markdown(original_title, unsafe_allow_html=True)
 edited_df = st.data_editor(newValuedf1,hide_index=True,use_container_width=True)
 
-pos = np.float(edited_df[['Hydrothermal Resource (positive)']].values[1])
+pos = edited_df[['Hydrothermal Resource (positive)']].values[1]
 #neg = float(edited_df[['No Hydrothermal Resource (negative)']].values[1])
 value_array, value_array_df = mymodule.make_value_array(count_ij, profit_drill_pos= pos, cost_drill_neg = -1e-6)
